Add cookie proxying
To be able to do things like log in a website, we intercept cookies from
our responses, store them in memory and re-send them to the website the
next time we hit it.

For now, the implementation is a bit messy because it involves fancy
inter-thread memory sharing and it's the first time I seriously dig in
that stuff. So for now, tons of `unwrap()` and `clone()` all around!

That's what I like about Rust: dangers and innefficiencies are always in
your face in the form of `unwrap()` and `clone()`. In another language,
code elegance could fool you into thinking it was solid code.